chore: Added authorisation to current user
