+++ src/App.css
@@ -6,19 +6,12 @@
 
 body {
   font-family: Arial, Helvetica, sans-serif;
-  /* line-height: 1.4; */
 }
 
-a {
-  color: #333;
-  text-decoration: none;
-}
 .App {
   background: #fff;
-  /* margin: 10px; */
 }
 .container {
-  /* padding: 0 1rem; */
   margin-left: 8px;
 }
