--- src/components/todos/Todos.js
@@ -1,37 +1,0 @@
-import React, { Component }  from 'react';
-import TodoItem from './TodoItem';
-import PropTypes from 'prop-types';
-
-class Todos extends Component {
-    render() {
-    return this.props.todos.map((todo) => (
-    <TodoItem key={todo.id} todo={todo} markComplete={this.props.markComplete} deleteTodo={this.props.deleteTodo} />
-    ));
-    }
-}
-
-// PropTypes
-Todos.propTypes = {
-    todos: PropTypes.array.isRequired,
-    markComplete: PropTypes.func.isRequired,
-    deleteTodo: PropTypes.func.isRequired
-}
-
-export default Todos;
-
-
-
-
-
-// import React, { Component }  from 'react';
-// import TodoItem from './TodoItem';
-
-// class Todos extends Component {
-//     render() {
-//     return this.props.todos.map((todo) => (
-//     <TodoItem key={todo.id} todo={todo}/>
-//     ));
-//     }
-// }
-
-// export default Todos;
